services/product-service/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk